[WEJBHTTP-139] Refactoring - moving ByteOutputs classes to common module
